+++ lib/register_sources_dk/services/es_index_creator.rb
@@ -9,21 +9,17 @@ module RegisterSourcesDk
     class EsIndexCreator
       MAPPINGS = JSON.parse(File.read(File.expand_path('mappings/vrdeltagerperson.json', __dir__)))
 
-      def initialize(
-        client: Config::ELASTICSEARCH_CLIENT,
-        index: Config::ELASTICSEARCH_INDEX
-      )
+      def initialize(client: Config::ELASTICSEARCH_CLIENT)
         @client = client
-        @index = index
       end
 
-      def create_index
+      def create_index(index)
         client.indices.create index:, body: { mappings: MAPPINGS }
       end
 
       private
 
-      attr_reader :client, :index
+      attr_reader :client
     end
   end
 end
