+++ spec/unit/services/es_index_creator_spec.rb
@@ -3,7 +3,7 @@
 require 'register_sources_dk/services/es_index_creator'
 
 RSpec.describe RegisterSourcesDk::Services::EsIndexCreator do
-  subject { described_class.new(client:, index:) }
+  subject { described_class.new(client:) }
 
   let(:client) { double 'client', indices: double('indices') }
   let(:index) { double 'index' }
@@ -14,7 +14,7 @@ RSpec.describe RegisterSourcesDk::Services::EsIndexCreator do
         index:
       )
 
-      subject.create_index
+      subject.create_index(index)
     end
   end
 end
